Split node-list into multiple components
