+++ src/components/node-list/node-list.scss
@@ -117,8 +117,9 @@
   background: none;
   border: none;
   box-shadow: none;
-  padding: 1em;
+  padding: 0.5em;
   margin: 0.3em;
+  font-size: 1.3em;
   color: inherit;
   font-family: inherit;
   cursor: pointer;
